Track fan's last running speed so it turns on at that speed
This fixes an issue where fan thermostat devices were reporting
their speed as "on" for a little while after being turned on.
This happened because the child device didn't know what speed the
fan would turn on to until it processed the controlled fan's speed
evnet.  This tracks the last running speed when the fan thermostat
device turns off to try to guess what speed the fan will turn on
at.