Add support for the "off" thermostat mode and SwitchLevel capability
This adds support for the "off" thermostat mode.  When in the "off"
mode all controlled fans will be turned off.  The manual override
timer will work the same as in the "cool" mode.

This also adds support fot the SwitchLevel capability.  The setLevel
command can be used to control the fan speed just like setSpeed.
The range is split into a number of equal-sized ranges that each
set the fan to a given speed.  For example, for a 3-speed fan,
level 1-33 will set the fan to low, 34-66 to medium and 67-99 high.
When the thermostat sets the fan speed it will set the level to the
bottom of the corresponding range.